feat(LIVE-9328): refactor hook
